Fix circular dependency and null pool stats
