+++ lifecycle/lifecycle/endpoints/health.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI
 from fastapi.responses import JSONResponse
 
 from lifecycle.config import Config
-from lifecycle.server.db_status import database_status
+from lifecycle.server.cache import LifecycleCache
 
 
 def setup_health_endpoint(api: FastAPI, config: Config):
@@ -28,13 +28,12 @@ def setup_health_endpoint(api: FastAPI, config: Config):
     @api.get("/health", tags=['root'])
     def _health():
         """Report current application status"""
-        ready = database_status.connected
-        status_code = 200 if ready else 500
+        database_connected = LifecycleCache.db_engine().database_status().connected
+        status_code = 200 if database_connected else 500
         content = {
             'service': 'lifecycle',
             'live': True,
-            'ready': ready,
-            'database_connected': database_status.connected,
+            'database_connected': database_connected,
             'git_version': os.environ.get('GIT_VERSION', 'dev'),
             'docker_tag': os.environ.get('DOCKER_TAG', ''),
             'auth_required': config.auth_required,
