+++ lifecycle/tests/server/test_version.py
@@ -3,7 +3,7 @@ import os
 from fastapi.testclient import TestClient
 
 from lifecycle.config import Config
-from lifecycle.server.metrics import unregister_metrics
+from lifecycle.server.metrics_collector import unregister_metrics
 from racetrack_commons.plugin.engine import PluginEngine
 from lifecycle.server.api import create_fastapi_app
 
@@ -21,4 +21,4 @@ def test_health_version_endpoint():
     obj = response.json()
     assert obj['git_version'] == '0.0.1-g32c4b29-dirty'
     assert obj['live'] is True
-    assert obj['ready'] is True
+
